# docs/openapi_complete.yaml
openapi: 3.0.3
info:
  title: AI Secretary SaaS API
  description: |
    Comprehensive API for the AI Secretary SaaS platform - an omnichannel solution for SMBs that provides 
    multi-agent AI secretary system with Inbox management, CRM functionality, Calendar scheduling, 
    RAG-based knowledge management, Stripe invoicing, and KYB counterparty monitoring.
    
    ## Features
    - Multi-channel communication (Telegram, Signal, Web widget)
    - AI-powered message routing and responses with specialized agents
    - CRM with lead management, pipeline tracking, tasks, and notes
    - Google Calendar integration with appointment booking
    - Knowledge base with RAG search and document processing
    - Stripe billing and subscription management with usage metering
    - KYB monitoring with external data sources (VIES, GLEIF, sanctions)
    - Multi-tenant architecture with role-based access control
    - GDPR compliance features with data retention and deletion
    - Real-time WebSocket communication
    - Comprehensive monitoring and analytics
    
    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    To obtain tokens, use the `/auth/login` endpoint with your credentials.
    
    ## Rate Limiting
    API requests are rate limited per tenant and user. Rate limit headers are included in responses:
...
    
    ## Pagination
...
    
...
    
...
    {
...
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register new tenant
      description: Register new tenant and owner user with 3-day trial
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - organization_name
              properties:
                email:
                  type: string
                  format: email
                  example: "owner@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "securepassword123"
                organization_name:
                  type: string
                  minLength: 2
                  example: "Example Corp"
                first_name:
                  type: string
                  example: "John"
                last_name:
                  type: string
                  example: "Doe"
                language:
                  type: string
                  enum: [en, de, uk]
                  default: en
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      tenant:
                        $ref: '#/components/schemas/Tenant'
                      access_token:
                        type: string
                      refresh_token:
                        type: string
                      token_type:
                        type: string
                        example: "Bearer"
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticate user and return JWT tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "securepassword123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      tenant:
                        $ref: '#/components/schemas/Tenant'
                      access_token:
                        type: string
                        example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                      refresh_token:
                        type: string
                        example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                      token_type:
                        type: string
                        example: "Bearer"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    RefreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Conflict:
      description: Resource already exists or conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
              example: "req_123456789"
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [owner, manager, support, accounting, read_only]
        is_active:
          type: boolean
        language:
          type: string
          enum: [en, de, uk]
        timezone:
          type: string
        last_login:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Tenant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        subscription_status:
          type: string
          enum: [trial, active, past_due, canceled, unpaid]
        trial_ends_at:
          type: string
          format: date-time
          nullable: true
        settings:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...